Moved estimation to a separate package
