+++ tests/surrogate_tests/test_causal_surrogate_assisted.py
@@ -4,20 +4,25 @@ from causal_testing.specification.causal_dag import CausalDAG
 from causal_testing.specification.causal_specification import CausalSpecification
 from causal_testing.specification.scenario import Scenario
 from causal_testing.specification.variable import Input, Output
-from causal_testing.surrogate.causal_surrogate_assisted import SimulationResult, CausalSurrogateAssistedTestCase, Simulator
+from causal_testing.surrogate.causal_surrogate_assisted import (
+    SimulationResult,
+    CausalSurrogateAssistedTestCase,
+    Simulator,
+)
 from causal_testing.surrogate.surrogate_search_algorithms import GeneticSearchAlgorithm
-from causal_testing.testing.estimators import CubicSplineRegressionEstimator
+from causal_testing.estimation.cubic_spline_estimator import CubicSplineRegressionEstimator
 
 import os
 import shutil, tempfile
 import pandas as pd
 import numpy as np
 
+
 class TestSimulationResult(unittest.TestCase):
 
     def setUp(self):
 
-        self.data = {'key': 'value'}
+        self.data = {"key": "value"}
 
     def test_inputs(self):
 
@@ -36,6 +41,7 @@ class TestSimulationResult(unittest.TestCase):
                     self.assertEqual(result.fault, fault)
 
                     self.assertEqual(result.relationship, relationship)
+
 
 class TestCausalSurrogate(unittest.TestCase):
 
@@ -79,20 +85,18 @@ class TestCausalSurrogate(unittest.TestCase):
         x = Input("X", float)
         m = Input("M", int)
         y = Output("Y", float)
-        scenario = Scenario(variables={z, x, m, y}, constraints={
-            z <= 0, z >= 3,
-            x <= 0, x >= 3,
-            m <= 0, m >= 3
-        })
+        scenario = Scenario(variables={z, x, m, y}, constraints={z <= 0, z >= 3, x <= 0, x >= 3, m <= 0, m >= 3})
         specification = CausalSpecification(scenario, causal_dag)
 
-        search_algorithm = GeneticSearchAlgorithm(config= {
+        search_algorithm = GeneticSearchAlgorithm(
+            config={
                 "parent_selection_type": "tournament",
                 "K_tournament": 4,
                 "mutation_type": "random",
                 "mutation_percent_genes": 50,
                 "mutation_by_replacement": True,
-            })
+            }
+        )
         simulator = TestSimulator()
 
         c_s_a_test_case = CausalSurrogateAssistedTestCase(specification, search_algorithm, simulator)
@@ -111,20 +115,18 @@ class TestCausalSurrogate(unittest.TestCase):
         x = Input("X", float)
         m = Input("M", int)
         y = Output("Y", float)
-        scenario = Scenario(variables={z, x, m, y}, constraints={
-            z <= 0, z >= 3,
-            x <= 0, x >= 3,
-            m <= 0, m >= 3
-        })
+        scenario = Scenario(variables={z, x, m, y}, constraints={z <= 0, z >= 3, x <= 0, x >= 3, m <= 0, m >= 3})
         specification = CausalSpecification(scenario, causal_dag)
 
-        search_algorithm = GeneticSearchAlgorithm(config= {
+        search_algorithm = GeneticSearchAlgorithm(
+            config={
                 "parent_selection_type": "tournament",
                 "K_tournament": 4,
                 "mutation_type": "random",
                 "mutation_percent_genes": 50,
                 "mutation_by_replacement": True,
-            })
+            }
+        )
         simulator = TestSimulatorFailing()
 
         c_s_a_test_case = CausalSurrogateAssistedTestCase(specification, search_algorithm, simulator)
@@ -143,26 +145,25 @@ class TestCausalSurrogate(unittest.TestCase):
         x = Input("X", float)
         m = Input("M", int)
         y = Output("Y", float)
-        scenario = Scenario(variables={z, x, m, y}, constraints={
-            z <= 0, z >= 3,
-            x <= 0, x >= 3,
-            m <= 0, m >= 3
-        })
+        scenario = Scenario(variables={z, x, m, y}, constraints={z <= 0, z >= 3, x <= 0, x >= 3, m <= 0, m >= 3})
         specification = CausalSpecification(scenario, causal_dag)
 
-        search_algorithm = GeneticSearchAlgorithm(config= {
+        search_algorithm = GeneticSearchAlgorithm(
+            config={
                 "parent_selection_type": "tournament",
                 "K_tournament": 4,
                 "mutation_type": "random",
                 "mutation_percent_genes": 50,
                 "mutation_by_replacement": True,
-            })
+            }
+        )
         simulator = TestSimulator()
 
         c_s_a_test_case = CausalSurrogateAssistedTestCase(specification, search_algorithm, simulator)
 
-        result, iterations, result_data = c_s_a_test_case.execute(ObservationalDataCollector(scenario, df), 
-                                                                    custom_data_aggregator=data_double_aggregator)
+        result, iterations, result_data = c_s_a_test_case.execute(
+            ObservationalDataCollector(scenario, df), custom_data_aggregator=data_double_aggregator
+        )
 
         self.assertIsInstance(result, SimulationResult)
         self.assertEqual(iterations, 1)
@@ -176,62 +177,69 @@ class TestCausalSurrogate(unittest.TestCase):
         x = Input("X", float)
         m = Input("M", int)
         y = Output("Y", float)
-        scenario = Scenario(variables={z, x, m, y}, constraints={
-            z <= 0, z >= 3,
-            x <= 0, x >= 3,
-            m <= 0, m >= 3
-        })
+        scenario = Scenario(variables={z, x, m, y}, constraints={z <= 0, z >= 3, x <= 0, x >= 3, m <= 0, m >= 3})
         specification = CausalSpecification(scenario, causal_dag)
 
-        search_algorithm = GeneticSearchAlgorithm(config= {
+        search_algorithm = GeneticSearchAlgorithm(
+            config={
                 "parent_selection_type": "tournament",
                 "K_tournament": 4,
                 "mutation_type": "random",
                 "mutation_percent_genes": 50,
                 "mutation_by_replacement": True,
-                "gene_space": "Something"
-            })
+                "gene_space": "Something",
+            }
+        )
         simulator = TestSimulator()
 
         c_s_a_test_case = CausalSurrogateAssistedTestCase(specification, search_algorithm, simulator)
 
-        self.assertRaises(ValueError, c_s_a_test_case.execute, 
-                          data_collector=ObservationalDataCollector(scenario, df),
-                          custom_data_aggregator=data_double_aggregator)
+        self.assertRaises(
+            ValueError,
+            c_s_a_test_case.execute,
+            data_collector=ObservationalDataCollector(scenario, df),
+            custom_data_aggregator=data_double_aggregator,
+        )
 
     def tearDown(self) -> None:
         shutil.rmtree(self.temp_dir_path)
 
+
 def load_class_df():
     """Get the testing data and put into a dataframe."""
 
-    class_df = pd.DataFrame({"Z": np.arange(16), "X": np.arange(16), "M": np.arange(16, 32), "Y": np.arange(32,16,-1)})
+    class_df = pd.DataFrame(
+        {"Z": np.arange(16), "X": np.arange(16), "M": np.arange(16, 32), "Y": np.arange(32, 16, -1)}
+    )
     return class_df
+
 
 class TestSimulator(Simulator):
 
     def run_with_config(self, configuration: dict) -> SimulationResult:
         return SimulationResult({"Z": 1, "X": 1, "M": 1, "Y": 1}, True, None)
-    
+
     def startup(self):
         pass
 
     def shutdown(self):
         pass
+
 
 class TestSimulatorFailing(Simulator):
 
     def run_with_config(self, configuration: dict) -> SimulationResult:
         return SimulationResult({"Z": 1, "X": 1, "M": 1, "Y": 1}, False, None)
-    
+
     def startup(self):
         pass
 
     def shutdown(self):
         pass
 
+
 def data_double_aggregator(data, new_data):
-    """Previously used data.append(new_data), however, pandas version >2 requires pd.concat() since append is now a private method. 
-       Converting new_data to a pd.DataFrame is required to use pd.concat(). """
+    """Previously used data.append(new_data), however, pandas version >2 requires pd.concat() since append is now a private method.
+    Converting new_data to a pd.DataFrame is required to use pd.concat()."""
     new_data = pd.DataFrame([new_data])
     return pd.concat([data, new_data, new_data], ignore_index=True)
